+++ web/src/app/modules/views/main/editors/modules/graphical-editor/components/graphical-editor.component.ts
@@ -2,7 +2,6 @@ import { ChangeDetectionStrategy, Component, ElementRef, Input, OnDestroy, ViewC
 import { TranslateService } from '@ngx-translate/core';
 import { mxgraph } from 'mxgraph'; // Typings only - no code!
 import { Subscription } from 'rxjs';
-import { isFunction } from 'rxjs/internal-compatibility';
 import { CEGModel } from 'src/app/model/CEGModel';
 import { Process } from 'src/app/model/Process';
 import { ProcessConnection } from 'src/app/model/ProcessConnection';
@@ -34,6 +33,7 @@ import { EditorKeyHandler } from './editor-components/editor-key-handler';
 import { EditorPopup } from './editor-components/editor-popup';
 import { EditorStyle } from './editor-components/editor-style';
 import { GraphValidator } from './editor-components/graph-validator';
+import { CEGNodeTypeUtil } from './util/ceg-node-type-util';
 import { ChangeTranslator } from './util/change-translator';
 import { StyleChanger } from './util/style-changer';
 
@@ -489,7 +489,7 @@ export class GraphicalEditor implements OnDestroy {
             if (Type.is(this.model, CEGModel)) {
                 for (const url in vertexCache) {
                     const vertex = vertexCache[url];
-                    const type = GraphicalEditor.getCEGNodeType(vertex);
+                    const type = CEGNodeTypeUtil.getCEGNodeType(vertex);
                     StyleChanger.addStyle(vertex, this.graph, type);
                 }
             }
@@ -598,34 +598,6 @@ export class GraphicalEditor implements OnDestroy {
     /*********************** Editor Options ***********************/
     public get model(): CEGModel | Process {
         return this._model;
-    }
-
-    public static getCEGNodeType(cell: mxgraph.mxCell) {
-        if (cell.isEdge()) {
-            return '';
-        }
-
-        if (cell.edges === null) {
-            // Node without Edges
-            return EditorStyle.CAUSE_STYLE_NAME;
-        }
-
-        let hasIncommingEdges = false;
-        let hasOutgoingEdges = false;
-        for (const edge of cell.edges) {
-            if (edge.source.id === cell.id) {
-                hasOutgoingEdges = true;
-            } else if (edge.target.id === cell.id) {
-                hasIncommingEdges = true;
-            }
-        }
-
-        if (hasIncommingEdges && hasOutgoingEdges) {
-            return EditorStyle.INNER_STYLE_NAME;
-        } else if (hasIncommingEdges) {
-            return EditorStyle.EFFECT_STYLE_NAME;
-        }
-        return EditorStyle.CAUSE_STYLE_NAME;
     }
 
     private resetProviders(model: CEGModel | Process): void {
